Update Jillian Practice File.R
Tried to clean up the practice script. Created a basic graph of the NDVI data by day of the year, and tried to recreate the graph I remembered seeing while Ross was working with the script. Can y'all check that I was able to save them as png in the shared folder and give any suggestions? Let me know if anything else needs work. I'll definitely to write the notes as I work moving forward!! Thank you!