Checked AICc with April/May
Created linear models with only the months of April and May. Compared the AICc of the AprMay models with the Grow Season (April-September) models that performed better than the full year models. The full GS models performed better than any of the AprMay models, with the 60dSPI being the best fit.